requirements.txt
beautifulsoup4
requests
tinydb
huey
python_dotenv